requirements.txt
Flask==2.3.3
numpy==1.24.4
Pillow==10.0.1
dlib==19.24.0
opencv-python==4.8.1.78
face-recognition==1.3.0
Werkzeug==2.3.7
